Removed unneeded Files
