+++ src/components/MarketPlace/Selling/Selling.jsx
@@ -1,0 +1,8 @@
+function Selling() {
+  return (
+    <div className="selling">
+      <h2>Selling</h2>
+      <p>Manage your listings and sales here.</p>
+    </div>
+  );
+}
